backend/requirements.txt
Flask==2.3.3
pymongo==4.6.1
python-dotenv==1.0.0
flask-cors==4.0.0
PyJWT==2.8.0
flask-bcrypt==1.0.1
Werkzeug==2.3.7
python-dateutil==2.8.2
bson==0.5.10 